webapp/requirements.txt
langchain==0.0.350
streamlit-chat
transformers
torch
python-dotenv
accelerate
underthesea
sentence-transformers
boto3
requests
streamlit-feedback